requirements.txt
asgiref==3.6.0
click==8.1.6
colorama==0.4.6
contourpy==1.0.7
cssbeautifier==1.14.9
cycler==0.11.0
Django==4.1.7
djlint==1.32.1
EditorConfig==0.12.3
fonttools==4.39.4
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-resources==5.12.0
jsbeautifier==1.14.9
json5==0.9.14
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathspec==0.11.2
Pillow==9.5.0
PyMySQL==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
tzdata==2023.3
zipp==3.15.0
